canvas-sync/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
